Дописаны модули. Переписана валидация через объекты.
