+++ components/modal.js
@@ -1,0 +1,23 @@
+import { closePopup } from "./utils.js";
+import enableValidation from "./validate.js";
+import createCard from "./card.js";
+import { nameProfile, jobProfile, linkInput, titleInput, addPopup, elements, nameInput, jobInput, profilePopup, validationSettings } from "./index.js";
+
+// Функция подтверждения формы загрузки карточки
+function handleCardFormSubmit(evt) {
+  evt.preventDefault();
+  elements.prepend(createCard(linkInput.value, titleInput.value));
+  evt.target.reset();
+  enableValidation(validationSettings);
+  closePopup(addPopup);
+}
+
+// Функция настройки ввода формы
+function handleProfileFormSubmit(evt) {
+  evt.preventDefault();
+  nameProfile.textContent = nameInput.value;
+  jobProfile.textContent = jobInput.value;
+  closePopup(profilePopup);
+}
+
+export {handleCardFormSubmit, handleProfileFormSubmit};
