+++ index.html
@@ -72,6 +72,6 @@
       </div>
     </article>
   </template>
-  <script src="js/index.js"></script>
+  <script type="module" src="components/index.js"></script>
 </body>
 </html>
